pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "google-cloud-secret-manager>=2.23.3",
    "google-cloud-storage>=3.1.0",
    "googlemaps>=4.10.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]
